+++ Common.Services/IUnitConverter.cs
@@ -4,6 +4,6 @@ namespace Common.Services
 {
     public interface IUnitConverter
     {
-        public double Convert(double fromValue, string fromUnit, string toUnit, UnitType unitType);
+        public decimal Convert(decimal fromValue, string fromUnit, string toUnit, UnitType unitType);
     }
 }
